requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
Pillow==7.0.0
pyTelegramBotAPI==3.6.7
requests==2.7.0
six==1.14.0
urllib3==1.25.8
